refactor: report pub folder name
