+++ apps/web/src/pages/report/[id]/index.tsx
@@ -1,3 +1,3 @@
-import Report from '@components/Shared/Modal/Report';
+import Report from '@components/Shared/Modal/ReportPublication';
 
 export default Report;
